==> Part_3/Case_Study_2/1- assessing.ipynb <==
# Assessing
Use the space below to explore `all_alpha_08.csv` and `all_alpha_18.csv` to answer the quiz questions below.

import pandas as pd
df1 = pd.read_csv('all_alpha_08.csv')
df2 = pd.read_csv('all_alpha_18.csv')

df1.head()

df2.head()

df1.info()

df1.duplicated().sum()

df1.isnull().sum()

df2.info()

df2.duplicated().sum()

df2.isnull().sum()

df1.nunique()

df2.nunique()

pd.unique(df1['Fuel'])

pd.unique(df2['Fuel'])
